Update Traveling salesman.py
Im going to add a readme of what I did tomorrow! My computer about to die and so Im done for now.

All we are missing is to fix Victoria's mutate to make it work with the functions and then find a way to generate the next population.

Then we can have a loop or something that does everything